refactor(templates): remove `plugins/` directory in favor of inline configuration (its now much shorter)
