+++ lib/components/blocks/kanban/kanban.tsx
@@ -6,18 +6,18 @@ import { createPortal } from "preact/compat";
 import {
   Announcements,
   DndContext,
+  type DragEndEvent,
+  type DragOverEvent,
   DragOverlay,
+  type DragStartEvent,
   KeyboardSensor,
   MouseSensor,
   TouchSensor,
+  type UniqueIdentifier,
   useSensor,
   useSensors,
-  type DragEndEvent,
-  type DragOverEvent,
-  type DragStartEvent,
-  type UniqueIdentifier,
 } from "../../../deps/@dnd-kit/core.ts";
-import { SortableContext, arrayMove } from "../../../deps/@dnd-kit/sortable.ts";
+import { arrayMove, SortableContext } from "../../../deps/@dnd-kit/sortable.ts";
 import type { Table, UseTableOptions } from "../table/table.tsx";
 import { KanbanContainer } from "./kanban-container.tsx";
 import { coordinateGetter } from "./multiple-containers-keyboard-preset.ts";
