+++ lib/database.ts
@@ -1,5 +1,4 @@
-import { monotonicFactory } from "../../deps/ulid.ts";
-import { filterObjectsByKeyValues } from "./utils.ts";
+import { monotonicFactory } from "./deps/ulid.ts";
 
 export const ulid = monotonicFactory();
 
@@ -127,4 +126,17 @@ export function createDatabase(kv: Deno.Kv) {
     patch,
     remove,
   };
+}
+
+export function filterObjectsByKeyValues<T = unknown>(
+  data: T[],
+  filters: Record<string, any> = {},
+) {
+  // filter item out if any of the filters fail, otherwise keep it
+  return !Object.keys(filters).length ? data : data.filter((item) => {
+    return !Object.entries(filters).some(([key, value]) => {
+      const itemValue = _get(item, key, "").toString();
+      return itemValue?.toLowerCase() !== value?.toLowerCase(); // case insensitive
+    });
+  });
 }
