+++ lib/plugins/api/utils.ts
@@ -1,5 +1,3 @@
-import { _get } from "../../deps/lodash.get.ts";
-
 export const RESPONSES = {
   missingApiKey: () => new Response("Missing API key", { status: 401 }),
   invalidApiKey: () => new Response("Invalid API key", { status: 401 }),
@@ -49,17 +47,4 @@ export async function parseRequestBody(req: Request) {
       }
     }
   }
-}
-
-export function filterObjectsByKeyValues<T = unknown>(
-  data: T[],
-  filters: Record<string, any> = {},
-) {
-  // filter item out if any of the filters fail, otherwise keep it
-  return !Object.keys(filters).length ? data : data.filter((item) => {
-    return !Object.entries(filters).some(([key, value]) => {
-      const itemValue = _get(item, key, "").toString();
-      return itemValue?.toLowerCase() !== value?.toLowerCase(); // case insensitive
-    });
-  });
 }
