+++ lib/plugins/mod.ts
@@ -1,4 +1,4 @@
+export * from "./api/plugin.ts";
 export * from "./auth/plugin.ts";
 export * from "./cron/plugin.ts";
-export * from "./db/plugin.ts";
 export * from "./environments/plugin.ts";
