+++ templates/crm/islands/accounts.tsx
@@ -37,7 +37,9 @@ export function PageAccounts(props: { accounts: Account[] }) {
     columns: [
       {
         id: "actions",
-        cell: (props) => <TableRowActions {...props} endpoint="/api/accounts" />,
+        cell: (props) => (
+          <TableRowActions {...props} endpoint="/api/accounts" />
+        ),
       },
       {
         accessorKey: "name",
@@ -126,7 +128,9 @@ export function PageAccounts(props: { accounts: Account[] }) {
             other: "mdi-link",
           } as const;
           const items = Object.entries(links)
-            .filter(([name, value]) => value)
+            .filter(([name, value]) =>
+              value
+            )
             .map(([name, value]) => ({ name, value, className: ICONS[name] }));
           return (
             <div className="flex gap-2">
