+++ templates/crm/islands/contact.tsx
@@ -27,8 +27,8 @@ import {
   FormLabel,
   FormMessage,
   useForm,
-  zodResolver,
   type UseFormReturn,
+  zodResolver,
 } from "netzo/components/form.tsx";
 import { Input } from "netzo/components/input.tsx";
 import { Textarea } from "netzo/components/textarea.tsx";
