+++ templates/crm/islands/contacts.tsx
@@ -17,7 +17,7 @@ import {
 } from "netzo/components/blocks/table/table.tsx";
 import { Button } from "netzo/components/button.tsx";
 import { IconCopy } from "netzo/components/icon-copy.tsx";
-import { getContact, type Contact } from "../data/contacts.ts";
+import { type Contact, getContact } from "../data/contacts.ts";
 import { I18N, toDateTime } from "../data/mod.ts";
 
 export function PageContacts(props: { contacts: Contact[] }) {
@@ -50,7 +50,9 @@ export function PageContacts(props: { contacts: Contact[] }) {
     columns: [
       {
         id: "actions",
-        cell: (props) => <TableRowActions {...props} endpoint="/api/contacts" />,
+        cell: (props) => (
+          <TableRowActions {...props} endpoint="/api/contacts" />
+        ),
       },
       {
         accessorKey: "name",
@@ -102,7 +104,8 @@ export function PageContacts(props: { contacts: Contact[] }) {
             </div>
           );
         },
-        filterFn: (row, id, value) => value.includes(row.getValue(id)),
+        filterFn: (row, id, value) =>
+          value.includes(row.getValue(id)),
       },
       {
         accessorKey: "activities",
