+++ templates/crm/islands/deal.tsx
@@ -22,8 +22,8 @@ import {
   FormLabel,
   FormMessage,
   useForm,
-  zodResolver,
   type UseFormReturn,
+  zodResolver,
 } from "netzo/components/form.tsx";
 import { Input } from "netzo/components/input.tsx";
 import { Textarea } from "netzo/components/textarea.tsx";
