+++ templates/crm/routes/accounts/[id].tsx
@@ -2,7 +2,7 @@ import { defineRoute } from "$fresh/server.ts";
 import type { Account } from "../../data/accounts.ts";
 import type { Deal } from "../../data/deals.ts";
 import { PageAccount } from "../../islands/account.tsx";
-import { db } from "../../plugins/db.config.ts";
+import { db } from "../../netzo.config.ts";
 
 export default defineRoute(async (req, ctx) => {
   const { id } = ctx.params;
