+++ templates/crm/routes/accounts/index.tsx
@@ -1,7 +1,7 @@
 import { defineRoute } from "$fresh/server.ts";
 import type { Account } from "../../data/accounts.ts";
 import { PageAccounts } from "../../islands/accounts.tsx";
-import { db } from "../../plugins/db.config.ts";
+import { db } from "../../netzo.config.ts";
 
 // NOTE: cannot pass functions as props from routes (server) to islands (client)
 export default defineRoute(async (req, ctx) => {
