+++ templates/minimal/netzo.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig } from "netzo/mod.ts";
 import { unocss } from "netzo/plugins/unocss/plugin.ts";
-import unocssConfig from "./plugins/unocss.config.ts";
+import unocssConfig from "./unocss.config.ts";
 
 export default defineConfig({
   plugins: [
